+++ tests/testthat.R
@@ -1,5 +1,5 @@
 library(testthat)
 library(formatR)
 
-test_check("Rclean")
+test_check("provClean")